Moved pauseScreen.kt to the screen folder
